Update index.html
Hello, congratulations for the beautiful work.  I'm working on your "sinric" project to make a home thermostat.  Together with sinric I am also using blynk to create an app on my smartphone.  I have a question to ask you.  Since in esp32 I go to check the temperature read by the probe and if this is lower than the one set I turn on a relay.  Is it possible to review feedback on the relay status on alexa?  It would be nice to see the thermostat status (on / off) and (hot / cold).  thanks a lot